Fix frontend API calls

+++ src/main/java/cn/edu/sdu/orz/bug/service/ModuleService.java
@@ -1,23 +1,16 @@
 package cn.edu.sdu.orz.bug.service;
 
-import cn.edu.sdu.orz.bug.dto.ModuleDTO;
 import cn.edu.sdu.orz.bug.entity.Module;
-import cn.edu.sdu.orz.bug.entity.Project;
 import cn.edu.sdu.orz.bug.repository.ModuleRepository;
-import cn.edu.sdu.orz.bug.repository.ProjectRepository;
 import cn.edu.sdu.orz.bug.utils.Utils;
 import cn.edu.sdu.orz.bug.vo.ModuleCreateVO;
-import cn.edu.sdu.orz.bug.vo.ModuleQueryVO;
 import cn.edu.sdu.orz.bug.vo.ModuleUpdateVO;
-import cn.edu.sdu.orz.bug.vo.ModuleVO;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import java.util.NoSuchElementException;
-import java.util.function.Function;
 
 @Service
 public class ModuleService {
@@ -26,45 +19,15 @@ public class ModuleService {
     private ModuleRepository moduleRepository;
 
     @Autowired
-    private UserService userService;
-    @Autowired
-    private ProjectRepository projectRepository;
-    @Autowired
     private ProjectService projectService;
 
-    public String save(ModuleVO vO) {
-        Module bean = new Module();
-        BeanUtils.copyProperties(vO, bean);
-        bean = moduleRepository.save(bean);
-        return bean.getId();
-    }
-
-    public void delete(String id) {
-        moduleRepository.deleteById(id);
-    }
-
-    public void update(String id, ModuleUpdateVO vO) {
-        Module bean = requireOne(id);
-        BeanUtils.copyProperties(vO, bean);
-        moduleRepository.save(bean);
-    }
-
-    public ModuleDTO getById(String id) {
-        Module original = requireOne(id);
-        return toDTO(original);
-    }
+    @Autowired
+    private UserService userService;
 
     public boolean create(ModuleCreateVO moduleCreateVO, HttpSession httpSession) {
-//        User user = userService.getLoggedInUser(httpSession);
-//        if (user == null) {
-//            return false;
-//        }
-//        if (userService.isLoggedInUserNotAdmin(httpSession)) {
-//            Project project = requireOne(id);
-//            if (!project.getOwner().equals(user.getId())) {
-//                return false;
-//            }
-//        }
+        if (userService.isNotLoggedIn(httpSession)) {
+            return false;
+        }
         try {
             Module bean = new Module();
             BeanUtils.copyProperties(moduleCreateVO, bean, Utils.getNullPropertyNames(moduleCreateVO));
@@ -79,16 +42,9 @@ public class ModuleService {
     }
 
     public boolean modify(String id, ModuleUpdateVO moduleUpdateVO, HttpSession httpSession) {
-//        User user = userService.getLoggedInUser(httpSession);
-//        if (user == null) {
-//            return false;
-//        }
-//        if (userService.isLoggedInUserNotAdmin(httpSession)) {
-//            Project project = requireOne(id);
-//            if (!project.getOwner().equals(user.getId())) {
-//                return false;
-//            }
-//        }
+        if (userService.isNotLoggedIn(httpSession)) {
+            return false;
+        }
         try {
             Module bean = requireOne(id);
             BeanUtils.copyProperties(moduleUpdateVO, bean, Utils.getNullPropertyNames(moduleUpdateVO));
@@ -100,16 +56,9 @@ public class ModuleService {
     }
 
     public boolean remove(String id, HttpSession httpSession) {
-//        User user = userService.getLoggedInUser(httpSession);
-//        if (user == null) {
-//            return false;
-//        }
-//        if (userService.isLoggedInUserNotAdmin(httpSession)) {
-//            Project project = requireOne(id);
-//            if (!project.getOwner().equals(user.getId())) {
-//                return false;
-//            }
-//        }
+        if (userService.isNotLoggedIn(httpSession)) {
+            return false;
+        }
         try {
             Module bean = requireOne(id);
             moduleRepository.delete(bean);
@@ -117,16 +66,6 @@ public class ModuleService {
             return false;
         }
         return true;
-    }
-
-    public Page<ModuleDTO> query(ModuleQueryVO vO) {
-        throw new UnsupportedOperationException();
-    }
-
-    private ModuleDTO toDTO(Module original) {
-        ModuleDTO bean = new ModuleDTO();
-        BeanUtils.copyProperties(original, bean);
-        return bean;
     }
 
     public Module requireOne(String id) {
